# Conflicts:
#	lib/cocoapods/installer/xcode/pods_project_generator/pod_target_installer.rb
#	spec/cocoapods-integration-specs

+++ lib/cocoapods/installer/xcode/pods_project_generator/target_installer.rb
@@ -118,7 +118,6 @@ module Pod
               path.dirname.mkpath
               result = generator.save_as(path)
             end
-            clean_support_files_temp_dir if support_files_temp_dir.exist?
             result
           end
 
@@ -131,7 +130,7 @@ module Pod
           # Remove temp file whose store .prefix/config/dummy file.
           #
           def clean_support_files_temp_dir
-            support_files_temp_dir.rmtree
+            support_files_temp_dir.rmtree if support_files_temp_dir.exist?
           end
 
           # @return [String] The temp file path to store temporary files.
@@ -169,9 +168,9 @@ module Pod
               update_changed_file(generator, path)
               add_file_to_support_group(path) if add_to_support_group
 
+              relative_path_string = path.relative_path_from(sandbox.root).to_s
               native_target.build_configurations.each do |c|
-                relative_path = path.relative_path_from(sandbox.root)
-                c.build_settings['INFOPLIST_FILE'] = relative_path.to_s
+                c.build_settings['INFOPLIST_FILE'] = relative_path_string
               end
             end
           end
@@ -185,16 +184,22 @@ module Pod
           # @return [void]
           #
           def create_module_map(native_target)
-            path = target.module_map_path
+            path = target.module_map_path_to_write
             UI.message "- Generating module map file at #{UI.path(path)}" do
               generator = Generator::ModuleMap.new(target)
               yield generator if block_given?
               update_changed_file(generator, path)
               add_file_to_support_group(path)
 
+              linked_path = target.module_map_path
+              if path != linked_path
+                linked_path.dirname.mkpath
+                FileUtils.ln_sf(path, linked_path)
+              end
+
+              relative_path_string = target.module_map_path.relative_path_from(sandbox.root).to_s
               native_target.build_configurations.each do |c|
-                relative_path = path.relative_path_from(sandbox.root)
-                c.build_settings['MODULEMAP_FILE'] = relative_path.to_s
+                c.build_settings['MODULEMAP_FILE'] = relative_path_string
               end
             end
           end
@@ -211,7 +216,7 @@ module Pod
           # @return [void]
           #
           def create_umbrella_header(native_target)
-            path = target.umbrella_header_path
+            path = target.umbrella_header_path_to_write
             UI.message "- Generating umbrella header at #{UI.path(path)}" do
               generator = Generator::UmbrellaHeader.new(target)
               yield generator if block_given?
@@ -220,10 +225,15 @@ module Pod
               # Add the file to the support group and the native target,
               # so it will been added to the header build phase
               file_ref = add_file_to_support_group(path)
-              native_target.add_file_references([file_ref])
+              build_file = native_target.headers_build_phase.add_file_reference(file_ref)
+
+              linked_path = target.umbrella_header_path
+              if path != linked_path
+                linked_path.dirname.mkpath
+                FileUtils.ln_sf(path, linked_path)
+              end
 
               acl = target.requires_frameworks? ? 'Public' : 'Project'
-              build_file = native_target.headers_build_phase.build_file(file_ref)
               build_file.settings ||= {}
               build_file.settings['ATTRIBUTES'] = [acl]
             end
